Extract react-cosmos-plugin-open-file plugin
