+++ packages/react-cosmos-plugin-open-file/.npmignore
@@ -1,0 +1,2 @@
+src
+tsconfig.build.json
