+++ packages/react-cosmos-plugin-open-file/index.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/index';
