+++ packages/react-cosmos-plugin-open-file/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "react-cosmos-plugin-open-file",
+  "version": "5.7.2",
+  "description": "React Cosmos plugin for opening source files",
+  "repository": "https://github.com/react-cosmos/react-cosmos/tree/main/packages/react-cosmos-plugin-open-file",
+  "license": "MIT",
+  "main": "index.js",
+  "dependencies": {
+    "@skidding/launch-editor": "^2.2.3",
+    "open": "^8.4.0",
+    "react-cosmos-plugin": "^5.6.0",
+    "react-cosmos-shared2": "^5.7.1"
+  }
+}
