+++ packages/react-cosmos-plugin-open-file/src/@types/launch-editor.d.ts
