+++ packages/react-cosmos-plugin-open-file/src/index.ts
@@ -1,0 +1,1 @@
+export { openFileDevServerPlugin } from './openFile';
