+++ packages/react-cosmos-plugin-open-file/src/openFile.ts
@@ -3,8 +3,8 @@ import express from 'express';
 import fs from 'fs';
 import open from 'open';
 import path from 'path';
-import { CosmosConfig } from '../config/shared';
-import { DevServerPluginArgs } from '../shared/types';
+import { DevServerPluginArgs } from 'react-cosmos-plugin';
+import { CosmosConfig } from 'react-cosmos-shared2/cosmosConfig';
 
 type ReqQuery = { filePath: void | string; line: number; column: number };
