+++ packages/react-cosmos-plugin/src/index.ts
@@ -1,4 +1,2 @@
-export {
-  CosmosPluginConfig,
-  getCosmosPluginConfigs,
-} from './getCosmosPluginConfigs';
+export * from './getCosmosPluginConfigs';
+export * from './serverPluginTypes';
