+++ packages/react-cosmos-shared2/cosmosConfig.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/cosmosConfig';
