+++ packages/react-cosmos-shared2/cosmosConfig.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/cosmosConfig');
