+++ packages/react-cosmos/package.json
@@ -6,7 +6,6 @@
   "license": "MIT",
   "main": "index.js",
   "dependencies": {
-    "@skidding/launch-editor": "^2.2.3",
     "@skidding/webpack-hot-middleware": "^2.25.1",
     "chokidar": "^3.5.2",
     "core-js": "^3.20.3",
@@ -17,7 +16,6 @@
     "import-from": "^3.0.0",
     "lodash": "^4.17.21",
     "micromatch": "^4.0.4",
-    "open": "^8.4.0",
     "pem": "^1.14.4",
     "pkg-up": "^3.1.0",
     "react-cosmos-playground2": "^5.7.2",
