+++ packages/react-cosmos/src/plugins/webpack/cosmosConfig/createWebpackCosmosConfig.ts
@@ -1,6 +1,6 @@
 import path from 'path';
+import { CosmosConfig } from 'react-cosmos-shared2/cosmosConfig';
 import { resolveModule } from '../../../config/resolveModule';
-import { CosmosConfig } from '../../../config/shared';
 import { fileExists } from '../../../shared/fs';
 
 type WebpackCosmosConfig = {
