+++ packages/react-cosmos/src/shared/playgroundHtml.ts
@@ -4,11 +4,10 @@ import {
   PlaygroundConfig,
   PlaygroundMountArgs,
 } from 'react-cosmos-playground2';
-import { CosmosPluginConfig } from 'react-cosmos-plugin';
+import { CosmosPluginConfig, PlatformType } from 'react-cosmos-plugin';
+import { CosmosConfig } from 'react-cosmos-shared2/cosmosConfig';
 import url from 'url';
-import { CosmosConfig } from '../config/shared';
 import { getStaticPath } from './static';
-import { PlatformType } from './types';
 import { replaceKeys } from './utils';
 
 export const RENDERER_FILENAME = '_renderer.html';
