+++ packages/react-cosmos/src/shared/pluginConfigs.ts
@@ -1,5 +1,5 @@
 import { getCosmosPluginConfigs } from 'react-cosmos-plugin';
-import { CosmosConfig } from '../config/shared';
+import { CosmosConfig } from 'react-cosmos-shared2/cosmosConfig';
 
 export function getPluginConfigs(cosmosConfig: CosmosConfig) {
   const { rootDir, disablePlugins, exportPath } = cosmosConfig;
