+++ packages/react-cosmos/src/userDeps/generateUserDepsModule.ts
@@ -1,4 +1,4 @@
-import { CosmosConfig } from '../config/shared';
+import { CosmosConfig } from 'react-cosmos-shared2/cosmosConfig';
 import { findUserModulePaths } from './findUserModulePaths';
 import { Json } from './shared';
 import { userDepsTemplate } from './userDepsTemplate';
